Moved pattern and request abstractions.
