+++ src/Glitter/Patterns/Behavioral/ICommand.cs
@@ -14,29 +14,24 @@
    limitations under the License.
 */
 
-namespace Glitter.Behaviors;
+namespace Glitter.Patterns.Behavioral;
 
 /// <summary>
-/// Represents a mediator.
+/// Represents a command.
 /// </summary>
-public interface IMediator
+public interface ICommand
 {
     /// <summary>
-    /// Sends the specified request.
+    /// Executes the command.
     /// </summary>
-    /// <typeparam name="TRequest">Specifies the type of the request.</typeparam>
-    /// <param name="request">The request to send.</param>
     /// <param name="cancellationToken">A cancellation token to support cancellation of the request.</param>
     /// <returns>A task that represents the asynchronous operation.</returns>
-    Task Send<TRequest>(IRequest request, CancellationToken cancellationToken = default)
-        where TRequest : IRequest;
+    Task Execute(CancellationToken cancellationToken = default);
 
     /// <summary>
-    /// Sends the specified request.
+    /// Rolls back the command.
     /// </summary>
-    /// <typeparam name="TResponse">Specifies the type of the response.</typeparam>
-    /// <param name="request">The request to send.</param>
     /// <param name="cancellationToken">A cancellation token to support cancellation of the request.</param>
-    /// <returns>The result of the request.</returns>
-    Task<TResponse> Send<TResponse>(IRequest<TResponse> request, CancellationToken cancellationToken = default);
+    /// <returns>A task that represents the asynchronous operation.</returns>
+    Task Rollback(CancellationToken cancellationToken = default);
 }
